requirements.txt
# Core ML frameworks
torch==2.8.0 --extra-index-url https://download.pytorch.org/whl/cpu
lightning>=2.0.0
torchmetrics>=1.0.0

# Data processing and analysis
numpy>=1.21.0
pandas>=2.0.0
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.5.0
plotly>=5.0.0

# Jupyter environment
ipywidgets>=8.0.0

# Configuration
PyYAML>=6.0.0

# Progress bars
tqdm>=4.0.0

# model summary
torchsummary>=1.5.0